requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.9.1
django-environ==0.4.5
django-redis
django-rest-framework==0.1.0
djangorestframework==3.12.4
PyJWT==2.3.0
djangorestframework-simplejwt==5.0.0
drf-yasg
python-dotenv==0.19.1
pytest
pytest-django==4.1.0
mock==4.0.3
factory-boy==3.2.0
openexchangerate==1.5.5

attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
distro==1.6.0
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
paramiko==2.8.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
texttable==1.6.4
urllib3==1.26.7
websocket-client==0.59.0
